requirements.txt
streamlit
openai
python-dotenv
audio-recorder-streamlit
gTTS
gptrim
duckduckgo_search
qdrant-client
jupyter
ipykernel
numpy
pandas
matplotlib
tiktoken
langchain
opencv-python
elevenlabs